Refactor ReferenceReplaceStep
